// src/models/user/IStrangerUser/IStrangerUserPreview.ts
export interface IStrangerUserPreview {
    Username: string;
    Profilid: string;
    Birthday: string;
    Zip: number;
    City: string;
    Picture: string;
    ProfilPicture?: string;
    Date?: string;
    Unlocked?: boolean;
    IsLiked?: boolean;
    CreateDate: string;
    Verified: number;
    IsOnline: boolean;
}